+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Impl `Eq` for `Bytea`, `Jsonb` and `Hstore`;
 - Impl `Clone` for `inspect::*` types;
 - Impl `FromSql`/`ToSql` for `[u8; N]`;
+- Impl `TryFrom<elephantry::Interval>` for `std::time::Duration`;
 
 ### Fixed
 - `u8` conversion;